requirements.txt
Django>=1.3.1
mimeparse>=0.1.3
python-dateutil==1.5
django-tastypie==0.9.11
pystache==0.3.1
